+++ src/base/net/stratum/Pools.h
@@ -72,7 +72,7 @@ public:
     inline bool operator==(const Pools &other) const    { return isEqual(other); }
 
     bool isEqual(const Pools &other) const;
-    int donateLevel() const;
+    float donateLevel() const;
     IStrategy *createStrategy(IStrategyListener *listener) const;
     rapidjson::Value toJSON(rapidjson::Document &doc) const;
     size_t active() const;
@@ -82,12 +82,12 @@ public:
     void toJSON(rapidjson::Value &out, rapidjson::Document &doc) const;
 
 private:
-    void setDonateLevel(int level);
+    void setDonateLevel(float level);
     void setProxyDonate(int value);
     void setRetries(int retries);
     void setRetryPause(int retryPause);
 
-    int m_donateLevel;
+    float m_donateLevel;
     int m_retries               = 5;
     int m_retryPause            = 5;
     ProxyDonate m_proxyDonate   = PROXY_DONATE_AUTO;
